refactor: create directories for hooks
