+++ src/components/Button/Button.test.tsx
@@ -1,6 +1,7 @@
-import { shallow } from 'enzyme';
+import { configure, shallow } from 'enzyme';
 import { Button } from './Button';
-
+import Adapter from 'enzyme-adapter-react-16';
+configure({ adapter: new Adapter() });
 describe('components/Button', () => {
   it('should have primary style render', () => {
     expect(shallow(<Button styleType="primary" />)).toMatchSnapshot();
