+++ src/components/Submenu/SubmenuSubItem/SubmenuSubItem.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import classNames from 'classnames';
 import { Link } from 'react-router-dom';
-import { useQueryParams } from '../../../hooks/useQueryParams';
+import { useQueryParams } from '../../../hooks/useQueryParams/useQueryParams';
 
 interface SubmenuSubItemProps extends React.HTMLAttributes<HTMLElement> {
   href: string;
